Moved LazyOption into separate folder
