+++ near-sdk/src/store/lazy/mod.rs
@@ -5,9 +5,6 @@
 //! [`Lazy`] and it will not be deserialized until requested.
 
 mod impls;
-mod lazy_option;
-
-pub use lazy_option::LazyOption;
 
 use borsh::{BorshDeserialize, BorshSerialize};
 use once_cell::unsync::OnceCell;
@@ -30,7 +27,7 @@ fn expect_consistent_state<T>(val: Option<T>) -> T {
     val.unwrap_or_else(|| env::panic(ERR_DELETED))
 }
 
-fn load_and_deserialize<T>(key: &[u8]) -> CacheEntry<T>
+pub(crate) fn load_and_deserialize<T>(key: &[u8]) -> CacheEntry<T>
 where
     T: BorshDeserialize,
 {
@@ -39,7 +36,7 @@ where
     CacheEntry::new_cached(Some(val))
 }
 
-fn serialize_and_store<T>(key: &[u8], value: &T)
+pub(crate) fn serialize_and_store<T>(key: &[u8], value: &T)
 where
     T: BorshSerialize,
 {
