+++ near-sdk/src/store/lazy_option/mod.rs
@@ -1,8 +1,10 @@
+mod impls;
+
 use borsh::{BorshDeserialize, BorshSerialize};
 use once_cell::unsync::OnceCell;
 
-use super::{load_and_deserialize, serialize_and_store};
 use crate::env;
+use crate::store::lazy::{load_and_deserialize, serialize_and_store};
 use crate::utils::{CacheEntry, EntryState};
 use crate::IntoStorageKey;
