+++ near-sdk/src/store/mod.rs
@@ -1,2 +1,4 @@
 mod lazy;
-pub use lazy::{Lazy, LazyOption};
+mod lazy_option;
+pub use lazy::Lazy;
+pub use lazy_option::LazyOption;
